backend/requirements.txt
atomicwrites==1.3.0
attrs==19.2.0
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.6
django-configurations==2.1
django-cors-headers==3.1.1
djangorestframework==3.10.3
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.2.0
pytest-cov==2.7.1
pytest-django==3.5.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.6.0
